@@ -1,12 +1,12 @@
 module github.com/uptrace/bun/dialect/mssqldialect
 
-go 1.18
+go 1.19
 
 replace github.com/uptrace/bun => ../..
 
 require (
-	github.com/uptrace/bun v1.1.13
-	golang.org/x/mod v0.10.0
+	github.com/uptrace/bun v1.1.15
+	golang.org/x/mod v0.12.0
 )
 
 require (
@@ -14,5 +14,5 @@ require (
 	github.com/tmthrgd/go-hex v0.0.0-20190904060850-447a3041c3bc // indirect
 	github.com/vmihailenco/msgpack/v5 v5.3.5 // indirect
 	github.com/vmihailenco/tagparser/v2 v2.0.0 // indirect
-	golang.org/x/sys v0.8.0 // indirect
+	golang.org/x/sys v0.12.0 // indirect
 )
