@@ -2,5 +2,5 @@ package mssqldialect
 
 // Version is the current release version.
 func Version() string {
-	return "1.1.13"
+	return "1.1.15"
 }
