@@ -11,9 +11,7 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/metric"
-	"go.opentelemetry.io/otel/metric/global"
-	"go.opentelemetry.io/otel/metric/instrument"
-	semconv "go.opentelemetry.io/otel/semconv/v1.12.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.20.0"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/uptrace/bun"
@@ -22,20 +20,12 @@ import (
 	"github.com/uptrace/opentelemetry-go-extra/otelsql"
 )
 
-var (
-	tracer = otel.Tracer("github.com/uptrace/bun")
-	meter  = global.Meter("github.com/uptrace/bun")
-
-	queryHistogram, _ = meter.Int64Histogram(
-		"go.sql.query_timing",
-		instrument.WithDescription("Timing of processed queries"),
-		instrument.WithUnit("milliseconds"),
-	)
-)
-
 type QueryHook struct {
-	attrs         []attribute.KeyValue
-	formatQueries bool
+	attrs          []attribute.KeyValue
+	formatQueries  bool
+	tracer         trace.Tracer
+	meter          metric.Meter
+	queryHistogram metric.Int64Histogram
 }
 
 var _ bun.QueryHook = (*QueryHook)(nil)
@@ -45,6 +35,17 @@ func NewQueryHook(opts ...Option) *QueryHook {
 	for _, opt := range opts {
 		opt(h)
 	}
+	if h.tracer == nil {
+		h.tracer = otel.Tracer("github.com/uptrace/bun")
+	}
+	if h.meter == nil {
+		h.meter = otel.Meter("github.com/uptrace/bun")
+	}
+	h.queryHistogram, _ = h.meter.Int64Histogram(
+		"go.sql.query_timing",
+		metric.WithDescription("Timing of processed queries"),
+		metric.WithUnit("milliseconds"),
+	)
 	return h
 }
 
@@ -59,7 +60,7 @@ func (h *QueryHook) Init(db *bun.DB) {
 }
 
 func (h *QueryHook) BeforeQuery(ctx context.Context, event *bun.QueryEvent) context.Context {
-	ctx, _ = tracer.Start(ctx, "", trace.WithSpanKind(trace.SpanKindClient))
+	ctx, _ = h.tracer.Start(ctx, "", trace.WithSpanKind(trace.SpanKindClient))
 	return ctx
 }
 
@@ -77,7 +78,7 @@ func (h *QueryHook) AfterQuery(ctx context.Context, event *bun.QueryEvent) {
 	}
 
 	dur := time.Since(event.StartTime)
-	queryHistogram.Record(ctx, dur.Milliseconds(), metric.WithAttributes(labels...))
+	h.queryHistogram.Record(ctx, dur.Milliseconds(), metric.WithAttributes(labels...))
 
 	span := trace.SpanFromContext(ctx)
 	if !span.IsRecording() {
