@@ -81,7 +81,7 @@ func appendIn(fmter Formatter, b []byte, slice reflect.Value) []byte {
 	sliceLen := slice.Len()
 
 	if sliceLen == 0 {
-		return append(b, "NULL"...)
+		return dialect.AppendNull(b)
 	}
 
 	for i := 0; i < sliceLen; i++ {
@@ -103,4 +103,23 @@ func appendIn(fmter Formatter, b []byte, slice reflect.Value) []byte {
 		}
 	}
 	return b
+}
+
+//------------------------------------------------------------------------------
+
+func NullZero(value interface{}) QueryAppender {
+	return nullZero{
+		value: value,
+	}
+}
+
+type nullZero struct {
+	value interface{}
+}
+
+func (nz nullZero) AppendQuery(fmter Formatter, b []byte) (_ []byte, err error) {
+	if isZero(nz.value) {
+		return dialect.AppendNull(b), nil
+	}
+	return fmter.AppendValue(b, reflect.ValueOf(nz.value)), nil
 }
